Section recursion
Have attempted to now store all sections as part of a lot.
I have written a calculation for area of a polygon as a stopping point
for this (i.e. only recursing if larger than goal area). I have also
implemented a check for the closest intersection to the longestLine.
Currently nothing is drawing