Branch L-system
A more complex algorithm for road network's has been created. It uses a
starting vector as an axiom and the rule is that each axiom turns into a
new edge with the axiom as the starting point. It also has a rule for
how many branches a node can have